requirements.txt
cryptography
requests
pycryptodome